+++ CHANGELOG.md
@@ -4,7 +4,7 @@ CHANGELOG
 Unreleased
 ----------
 
-* add the suppression resource - search, find, isSuppressed and delete
+* add the suppression resource - search, find, isSuppressed, add and delete
 
 0.1.1 (2026-08-22)
 ------------------